==> notebooks/2_dataframe_merge.ipynb <==
import requests
from bs4 import BeautifulSoup
import json
import pandas as pd
import numpy as np
import re
import ast

# raw data 로드 및 shape 확인

product_shirt = pd.read_csv(f"{link}product_cate/001002_1page.csv")
product_hoodie = pd.read_csv(f"{link}product_cate/001004_1page.csv")
product_man2man = pd.read_csv(f"{link}product_cate/001005_1page.csv")
product_longs = pd.read_csv(f"{link}product_cate/001010_1page.csv")

product_shirt.shape, product_hoodie.shape, product_man2man.shape, product_longs.shape

review_shirt = pd.read_csv(f"{link}product_review/001002_review_1page.csv")
review_hoodie = pd.read_csv(f"{link}product_review/001004_review_1page.csv")
review_man2man = pd.read_csv(f"{link}product_review/001005_review_1page.csv")
review_longs = pd.read_csv(f"{link}product_review/001010_review_1page.csv")

review_shirt["cate"] = "shirt"
review_hoodie["cate"] = "hoodie"
review_man2man["cate"] = "man2man"
review_longs["cate"] = "longs"

review_shirt.shape, review_hoodie.shape, review_man2man.shape, review_longs.shape

size_shirt = pd.read_csv(f"{link}product_size/001002_size_1page.csv")
size_hoodie = pd.read_csv(f"{link}product_size/001004_size_1page.csv")
size_man2man = pd.read_csv(f"{link}product_size/001005_size_1page.csv")
size_longs = pd.read_csv(f"{link}product_size/001010_size_1page.csv")

size_shirt.shape, size_hoodie.shape, size_man2man.shape, size_longs.shape

# raw data 병합

# 상품데이터 병합
product_merge_df = pd.concat([product_shirt, product_hoodie, product_man2man, product_longs], axis = 0).reset_index(drop = True)
product_merge_df.shape

# 리뷰데이터 병합
review_merge_df = pd.concat([review_shirt, review_hoodie, review_man2man, review_longs], axis = 0).reset_index(drop = True)
review_merge_df.shape

# 사이즈 데이터 병합
size_merge_df = pd.concat([size_shirt, size_hoodie, size_man2man, size_longs], axis = 0).reset_index(drop = True)
size_merge_df.shape

# 리뷰 개수 300개가 안되는 상품

review_group_under_300 = review_merge_df.groupby("goodsNo").count().reset_index()

review_group_under_300[review_group_under_300["review"] != 300].shape

len(review_group_under_300[review_group_under_300["review"] != 300]["goodsNo"].tolist())

review_group_under_300_list = review_group_under_300[review_group_under_300["review"] != 300]["goodsNo"].tolist()

# 사이즈 데이터가 없는 상품

set(review_merge_df["goodsNo"].unique()) - set(size_merge_df["goodsNo"].unique())

len(set(review_merge_df["goodsNo"].unique()) - set(size_merge_df["goodsNo"].unique()))

size_no_information = list(set(review_merge_df["goodsNo"].unique()) - set(size_merge_df["goodsNo"].unique()))

# 최종적으로 drop 되어야 할 상품의 개수

set(review_group_under_300_list) & set(size_no_information)

drop_list_step1 = set(review_group_under_300_list + size_no_information)
len(drop_list_step1)

# step1 drop 처리 후 데이터

product_drop_df = product_merge_df[~product_merge_df["goodsNo"].isin(drop_list_step1)].reset_index(drop = True)
product_drop_df.shape

review_drop_df = review_merge_df[~review_merge_df["goodsNo"].isin(drop_list_step1)].reset_index(drop = True)
review_drop_df.shape

size_drop_df = size_merge_df[~size_merge_df["goodsNo"].isin(drop_list_step1)].reset_index(drop = True)
size_drop_df.shape

# step2 전처리 병합

# 리뷰 / 사이즈 데이터에 대해 정규식 적용
review_drop_df["buy_size"] = review_drop_df["buy_size"].apply(lambda x: re.sub(r'[가-힣()\[\]{}\/]', ' ', x))
size_drop_df["Size"] = size_drop_df["Size"].apply(lambda x: re.sub(r'[가-힣()\[\]{}\/]', ' ', x))

# 띄어쓰기 기준으로 검사를 위해 split
review_drop_df["buy_size"] = review_drop_df["buy_size"].apply(lambda x: x.strip().split(" "))
size_drop_df["Size"] = size_drop_df["Size"].apply(lambda x: x.strip().split(" "))

def clean_list(lst):
    # 빈 문자열과 'NO'를 제거
    cleaned_list = [item for item in lst if item not in ['', 'NO']]

    # 특정 문자열 조합을 결합
    combined_list = []
    i = 0
    while i < len(cleaned_list):
        if i < len(cleaned_list) - 1 and cleaned_list[i] == "EXTRA" and cleaned_list[i + 1] == "LARGE":
            combined_list.append("EXTRALARGE")
            i += 2  # 두 항목을 하나로 합쳤으므로 두 인덱스를 건너뜀
        else:
            combined_list.append(cleaned_list[i])
            i += 1  # 다음 인덱스로 이동

    return combined_list

review_drop_df["buy_size"] = review_drop_df["buy_size"].apply(lambda x: clean_list(x))
size_drop_df["Size"] = size_drop_df["Size"].apply(lambda x: clean_list(x))

# 표기는 다르나 사이즈는 동일한 경우 중복제거
size_drop_df["Size"] = size_drop_df["Size"].astype(str)
size_drop_df.drop_duplicates(subset = ["goodsNo", "Size", "총장", "어깨너비", "가슴단면", "소매길이"], inplace = True)
size_drop_df["Size"] = size_drop_df["Size"].apply(ast.literal_eval)

# 한 상품에 대하여 unique값을 사이즈 표의 unique와 비교하여 매칭이 되면 병합

merge_temp_list = []

for i in review_drop_df["goodsNo"].unique():
    review_goods = review_drop_df[review_drop_df["goodsNo"] == i]
    size_goods = size_drop_df[size_drop_df["goodsNo"] == i]

    for review_idx, review_row in review_goods.iterrows():
        for size_idx, size_row in size_goods.iterrows():
            # 교집합이 있을 경우 병합
            if len(set(review_row["buy_size"]) & set(size_row["Size"])) >= 1:
                merged_row = pd.concat([review_row, size_row], axis=0).reset_index(drop = True)
                merge_temp_list.append(merged_row)

review_col_list = review_drop_df.columns.tolist()
size_col_list = size_drop_df.columns.tolist()
new_col = review_col_list + size_col_list
merge_temp_df = pd.DataFrame(merge_temp_list).reset_index(drop = True)

merge_temp_df.columns = ['goodsNo', 'id_date', 'body_size', 'review', 'buy_size', 'type_class', 'cate',
                        'goodsNo2', 'Size', 'length', 'shoulder', 'chest', 'sleeve']

# 최종 데이터프레임 병합

# 개수가 300개인 상품만 골라서 최종 병합
merge_temp_cnt = merge_temp_df.groupby("goodsNo").count().reset_index()
merge_temp_cnt300_list = merge_temp_cnt[merge_temp_cnt["review"] == 300]["goodsNo"].tolist()
total_temp_df = merge_temp_df[merge_temp_df["goodsNo"].isin(merge_temp_cnt300_list)].reset_index(drop = True)

len(merge_temp_cnt["goodsNo"].unique())

len(total_temp_df["goodsNo"].unique())

final_merge_df = pd.merge(product_drop_df, total_temp_df, on = "goodsNo")

final_merge_df

# final_merge_df.to_csv(f"{link}final_merge_df.csv", index = False)
