==> notebooks/5_review_analysis.ipynb <==
# 데이터로드

# 데이터로드

import pandas as pd
df = pd.read_csv("/content/drive/MyDrive/datathon/data_데이터/df_size_check_test5_okt.csv")

df.shape

len(df['review'].unique())

new_df = df[['goodsNo', 'brandNameEng', 'goodsName', 'imageUrl', 'reviewScore',
       'review', 'typeClass', 'cate', 'size', 'length', 'shoulder', 'chest',
       'sleeve', 'id', 'date', 'grade', 'gender', 'height', 'weight',
       'cate_review']]

new_df = new_df[~new_df["cate_review"].isnull()]

new_df.shape

# 카테고리별 dict

## 1) total_df : categories_total

import pandas as pd

# 사전 단어
total = ['사이즈', '핏', '핏입니다',
      '핏나','핏입니','옷핏','핏더', '옷',
      '상의','폼', '사이즈', '전체','품도','몸통','상체',
      '핏이구','핏이었','핏임','핏입니','핏감', '핏이','크기', '치수', '핏감입니', '핏좋', '핏듀', '핏도좋',
      '통', '품은', '너비', '품으', '허리', '밑위', '체감', '골반', '겨드랑이', '허벅지',
         '체구', '골격', '덩치', '몸매', '체격', '몸집', '핏나와', '핏좋아', '핏이였', '핏예쁘', '핏감도좋', '핏감좋']

small = ['작습니다','짧았어요','짧습니다','짧으면','짧','작았어요','짧다고','쪼여서','작았으면',
      '작아도','작으면','작음','작아진','좁다고',
      '짧아요','짧게','짧다는','좁다는','작네','작기는','달라붙는','작은데','짧은데','짧음',
      '작긴','짧네요','작게','작아서','작은','짧고','짧아','좁은','타이트','짧았지만','붙어서','짧긴',
      '작은것','작아요','짧은','짧아서','짧지만','좁아','좁고','작네요','작아','쪼이는','좁아요','크롭느낌',
      '크롭하','작고','숏','작다고','작다','좁아서','작','크롭','미니','크롭해',
      '길었으면','컸으면','좁게','짧지' ,'작지','크롭이','핏되',
      '짝','좁', '작았고','작긴한데','짧긴하지만','짧아여','짧은게','작긴하지만','작으니','작았습니다','짧거나','좁긴'.
       '작으니까','작은듯','끼긴','짧은듯','작았는데','조이는','끼네요','붙고','달라붙어서','붙음','붙어요',
       '짧은것','짧은거','끼는것','끼는','작은게','작음에도','짧은것이','좁네요','길었어도']

new_big = ['커도','남아','롱','긴게','넓게','덮어줘서','가오리','오퍼핏','부해','크구요','길고요','어벙함','컸었는데','길지도','길긴하지만','크니까','가리고',
 '박시하게','컸어도','덮는데','컸던','오벼핏','덮습니다','크고','덮','벌룬','덮을','헐렁하고','길거나','오버핏','박시하지만','덮이고','길게','길구요','길거든요',
 '걷어서','박스큽니다이','컷습니','컸고','박시하네요','박스핏','헐렁','덮어요','크','큼직해서','오바핏','길기는','넓아','크길래','넓어','루즈클까봐','낭',
 '클거','컸음','넓음','길다는','큰편인덴','넓었으면','수선해서','박시','널럴하','짧았으면','커욤','오비핏','넉','덮어서','덮네요','길어요','크지만','어벙',
 '덮구','길긴','낙','길어용','길지','덮이는','큰데','큰거','널','커서','펑퍼짐','길어','넓어요','길었고','더크','덮히는','넓길었는데','긴데','박시하','넓긴',
 '크긴','긴데도','덥는','접어도','넓은','크진','컸는데','큼직하니','벌룬핏','긴듯','커용','클까','오버사이즈','깁니','큼지막해서','접으면','박시해','가려요','커여',
 '넓은데','크셔서','큽니다만','크신','길었습니다','덮었어요','덮어용','큼','크니깐','헐렁해요','덮은','긴하고','박시하니','길','크고요',
 '박시한','크거나','길긴한데','큰데도','길더라고요','크더라','버핏','펑퍼짐해는','크다','헐렁해서','루즈핏','컸어용','헐렁한','빡시','긴','길어여','길었어요',
 '커요','펑퍼짐해서','넓네요','크네요','크면','와이드','가려져서','크다는','덮히고','긴것','길다고','컸을','오버','남는데','넓어서','컸어요','크다고','크게',
 '깁니다','길더라구요','헐렁거리','크더라고요','큰지','길어서','접고','덮이수선해서','컸습니다','컸지만','걷어','큼직하게','길엇어','잡아먹힌','박시하고','박시한데',
 '남는','컸으나','컸다','크더라구요','덮고도','크니','길었지만','여유핏','넓고','큽니','크네','접어야','기네','큰듯','크실','큽니당','넓긴한데',
 '길긴하네요','박시핏','긴거','넓우','덮음','여유롭게','클','펑퍼짐한','큰','길고','헐렁하게','큰것','아방핏','흘러내리는','덮고','크지']

# 두개의 사전이 합치고 나서 진행
# 중복제거
# 'n+2'개의 최종 리스트가 생성됨
def sizedict_nodup(oneOfRawList):
    # print('중복제거 전:{}개'.format(len(oneOfRawList)))
    oneOfRawList = set(oneOfRawList)
    # print('중복제거 후:{}개'.format(len(oneOfRawList)))
    # print('-'*10)
    return list(oneOfRawList)

new_total     = sizedict_nodup(new_total)
small     = sizedict_nodup(new_small)
big       = sizedict_nodup(new_big)

# 리스트와 리스트 이름들을 합쳐서 딕셔너리 생성
categories_total = {
    'Total': new_total,
    'Small': new_small,
    'Big': new_big
}

categories_total

## 2) length_df : categories_new_length

import pandas as pd

# 사전 단어

new_length = ['길이','기장','총장','궁디','궁딩','엉덩이','밑','밑단','위아래','끝단',
             '끝', '세로', '아랫쪽', '하단', '아랫', '아래쪽', '아래']

new_small = ['작습니다','짧았어요','짧습니다','짧으면','짧','작았어요','짧다고','쪼여서','작았으면',
      '작아도','작으면','작음','작아진','좁다고',
      '짧아요','짧게','짧다는','좁다는','작네','작기는','달라붙는','작은데','짧은데','짧음',
      '작긴','짧네요','작게','작아서','작은','짧고','짧아','좁은','타이트','짧았지만','붙어서','짧긴',
      '작은것','작아요','짧은','짧아서','짧지만','좁아','좁고','작네요','작아','쪼이는','좁아요','크롭느낌',
      '크롭하','작고','숏','작다고','작다','좁아서','작','크롭','미니','크롭해',
      '길었으면','컸으면','좁게','짧지' ,'작지','크롭이','핏되',
      '짝','좁', '작았고','작긴한데','짧긴하지만','짧아여','짧은게','작긴하지만','작으니','작았습니다','짧거나','좁긴'
                 '작으니까','작은듯','끼긴','짧은듯','작았는데','조이는','끼네요','붙고','달라붙어서','붙음','붙어요',
                 '짧은것','짧은거','끼는것','끼는','작은게','작음에도','짧은것이','좁네요','길었어도']

new_big = ['커도','남아','롱','긴게','넓게','덮어줘서','가오리','오퍼핏','부해','크구요','길고요','어벙함','컸었는데','길지도','길긴하지만','크니까','가리고',
 '박시하게','컸어도','덮는데','컸던','오벼핏','덮습니다','크고','덮','벌룬','덮을','헐렁하고','길거나','오버핏','박시하지만','덮이고','길게','길구요','길거든요',
 '걷어서','박스큽니다이','컷습니','컸고','박시하네요','박스핏','헐렁','덮어요','크','큼직해서','오바핏','길기는','넓아','크길래','넓어','루즈클까봐','낭',
 '클거','컸음','넓음','길다는','큰편인덴','넓었으면','수선해서','박시','널럴하','짧았으면','커욤','오비핏','넉','덮어서','덮네요','길어요','크지만','어벙',
 '덮구','길긴','낙','길어용','길지','덮이는','큰데','큰거','널','커서','펑퍼짐','길어','넓어요','길었고','더크','덮히는','넓길었는데','긴데','박시하','넓긴',
 '크긴','긴데도','덥는','접어도','넓은','크진','컸는데','큼직하니','벌룬핏','긴듯','커용','클까','오버사이즈','깁니','큼지막해서','접으면','박시해','가려요','커여',
 '넓은데','크셔서','큽니다만','크신','길었습니다','덮었어요','덮어용','큼','크니깐','헐렁해요','덮은','긴하고','박시하니','길','크고요',
 '박시한','크거나','길긴한데','큰데도','길더라고요','크더라','버핏','펑퍼짐해는','크다','헐렁해서','루즈핏','컸어용','헐렁한','빡시','긴','길어여','길었어요',
 '커요','펑퍼짐해서','넓네요','크네요','크면','와이드','가려져서','크다는','덮히고','긴것','길다고','컸을','오버','남는데','넓어서','컸어요','크다고','크게',
 '깁니다','길더라구요','헐렁거리','크더라고요','큰지','길어서','접고','덮이수선해서','컸습니다','컸지만','걷어','큼직하게','길엇어','잡아먹힌','박시하고','박시한데',
 '남는','컸으나','컸다','크더라구요','덮고도','크니','길었지만','여유핏','넓고','큽니','크네','접어야','기네','큰듯','크실','큽니당','넓긴한데',
 '길긴하네요','박시핏','긴거','넓우','덮음','여유롭게','클','펑퍼짐한','큰','길고','헐렁하게','큰것','아방핏','흘러내리는','덮고','크지']

# 두개의 사전이 합치고 나서 진행
# 중복제거
# 'n+2'개의 최종 리스트가 생성됨
def sizedict_nodup(oneOfRawList):
    # print('중복제거 전:{}개'.format(len(oneOfRawList)))
    oneOfRawList = set(oneOfRawList)
    # print('중복제거 후:{}개'.format(len(oneOfRawList)))
    # print('-'*10)
    return list(oneOfRawList)

new_length  = sizedict_nodup(new_length)
new_small     = sizedict_nodup(new_small)
new_big       = sizedict_nodup(new_big)

# 리스트와 리스트 이름들을 합쳐서 딕셔너리 생성
categories_length = {
    'Length': new_length,
    'Small': new_small,
    'Big': new_big
}

## 3) shoulder_df : categories_shoulder

import pandas as pd

# 사전 단어
new_shoulder = ['어깨핏','어께','어깨','너비','바디','가로',
            '어깡', '등판', '둘레', '광배', '어좁', '오비핏']

new_small = ['작습니다','짧았어요','짧습니다','짧으면','짧','작았어요','짧다고','쪼여서','작았으면',
      '작아도','작으면','작음','작아진','좁다고',
      '짧아요','짧게','짧다는','좁다는','작네','작기는','달라붙는','작은데','짧은데','짧음',
      '작긴','짧네요','작게','작아서','작은','짧고','짧아','좁은','타이트','짧았지만','붙어서','짧긴',
      '작은것','작아요','짧은','짧아서','짧지만','좁아','좁고','작네요','작아','쪼이는','좁아요','크롭느낌',
      '크롭하','작고','숏','작다고','작다','좁아서','작','크롭','미니','크롭해',
      '길었으면','컸으면','좁게','짧지' ,'작지','크롭이','핏되',
      '짝','좁', '작았고','작긴한데','짧긴하지만','짧아여','짧은게','작긴하지만','작으니','작았습니다','짧거나','좁긴'
                 '작으니까','작은듯','끼긴','짧은듯','작았는데','조이는','끼네요','붙고','달라붙어서','붙음','붙어요',
                 '짧은것','짧은거','끼는것','끼는','작은게','작음에도','짧은것이','좁네요','길었어도']

new_big = ['커도','남아','롱','긴게','넓게','덮어줘서','가오리','오퍼핏','부해','크구요','길고요','어벙함','컸었는데','길지도','길긴하지만','크니까','가리고',
 '박시하게','컸어도','덮는데','컸던','오벼핏','덮습니다','크고','덮','벌룬','덮을','헐렁하고','길거나','오버핏','박시하지만','덮이고','길게','길구요','길거든요',
 '걷어서','박스큽니다이','컷습니','컸고','박시하네요','박스핏','헐렁','덮어요','크','큼직해서','오바핏','길기는','넓아','크길래','넓어','루즈클까봐','낭',
 '클거','컸음','넓음','길다는','큰편인덴','넓었으면','수선해서','박시','널럴하','짧았으면','커욤','오비핏','넉','덮어서','덮네요','길어요','크지만','어벙',
 '덮구','길긴','낙','길어용','길지','덮이는','큰데','큰거','널','커서','펑퍼짐','길어','넓어요','길었고','더크','덮히는','넓길었는데','긴데','박시하','넓긴',
 '크긴','긴데도','덥는','접어도','넓은','크진','컸는데','큼직하니','벌룬핏','긴듯','커용','클까','오버사이즈','깁니','큼지막해서','접으면','박시해','가려요','커여',
 '넓은데','크셔서','큽니다만','크신','길었습니다','덮었어요','덮어용','큼','크니깐','헐렁해요','덮은','긴하고','박시하니','길','크고요',
 '박시한','크거나','길긴한데','큰데도','길더라고요','크더라','버핏','펑퍼짐해는','크다','헐렁해서','루즈핏','컸어용','헐렁한','빡시','긴','길어여','길었어요',
 '커요','펑퍼짐해서','넓네요','크네요','크면','와이드','가려져서','크다는','덮히고','긴것','길다고','컸을','오버','남는데','넓어서','컸어요','크다고','크게',
 '깁니다','길더라구요','헐렁거리','크더라고요','큰지','길어서','접고','덮이수선해서','컸습니다','컸지만','걷어','큼직하게','길엇어','잡아먹힌','박시하고','박시한데',
 '남는','컸으나','컸다','크더라구요','덮고도','크니','길었지만','여유핏','넓고','큽니','크네','접어야','기네','큰듯','크실','큽니당','넓긴한데',
 '길긴하네요','박시핏','긴거','넓우','덮음','여유롭게','클','펑퍼짐한','큰','길고','헐렁하게','큰것','아방핏','흘러내리는','덮고','크지']

# 중복제거
def sizedict_nodup(oneOfRawList):
    oneOfRawList = set(oneOfRawList)
    return list(oneOfRawList)

new_shoulder  = sizedict_nodup(new_shoulder)
new_small     = sizedict_nodup(new_small)
new_big       = sizedict_nodup(new_big)

# 리스트와 리스트 이름들을 합쳐서 딕셔너리 생성
categories_shoulder = {
    'Shoulder': new_shoulder,
    'Small': new_small,
    'Big': new_big
}

## 4) chest_df : categories_chest

import pandas as pd

# 사전 단어
new_chest = ['가슴','바디','가로','넓이','통','둘레','몸집','가슴팍','몸통', '너비', '넓고', '넓으며', '쉐잎']

new_small = ['작습니다','짧았어요','짧습니다','짧으면','짧','작았어요','짧다고','쪼여서','작았으면',
      '작아도','작으면','작음','작아진','좁다고',
      '짧아요','짧게','짧다는','좁다는','작네','작기는','달라붙는','작은데','짧은데','짧음',
      '작긴','짧네요','작게','작아서','작은','짧고','짧아','좁은','타이트','짧았지만','붙어서','짧긴',
      '작은것','작아요','짧은','짧아서','짧지만','좁아','좁고','작네요','작아','쪼이는','좁아요','크롭느낌',
      '크롭하','작고','숏','작다고','작다','좁아서','작','크롭','미니','크롭해',
      '길었으면','컸으면','좁게','짧지' ,'작지','크롭이','핏되',
      '짝','좁', '작았고','작긴한데','짧긴하지만','짧아여','짧은게','작긴하지만','작으니','작았습니다','짧거나','좁긴'
                 '작으니까','작은듯','끼긴','짧은듯','작았는데','조이는','끼네요','붙고','달라붙어서','붙음','붙어요',
                 '짧은것','짧은거','끼는것','끼는','작은게','작음에도','짧은것이','좁네요','길었어도']

new_big = ['커도','남아','롱','긴게','넓게','덮어줘서','가오리','오퍼핏','부해','크구요','길고요','어벙함','컸었는데','길지도','길긴하지만','크니까','가리고',
 '박시하게','컸어도','덮는데','컸던','오벼핏','덮습니다','크고','덮','벌룬','덮을','헐렁하고','길거나','오버핏','박시하지만','덮이고','길게','길구요','길거든요',
 '걷어서','박스큽니다이','컷습니','컸고','박시하네요','박스핏','헐렁','덮어요','크','큼직해서','오바핏','길기는','넓아','크길래','넓어','루즈클까봐','낭',
 '클거','컸음','넓음','길다는','큰편인덴','넓었으면','수선해서','박시','널럴하','짧았으면','커욤','오비핏','넉','덮어서','덮네요','길어요','크지만','어벙',
 '덮구','길긴','낙','길어용','길지','덮이는','큰데','큰거','널','커서','펑퍼짐','길어','넓어요','길었고','더크','덮히는','넓길었는데','긴데','박시하','넓긴',
 '크긴','긴데도','덥는','접어도','넓은','크진','컸는데','큼직하니','벌룬핏','긴듯','커용','클까','오버사이즈','깁니','큼지막해서','접으면','박시해','가려요','커여',
 '넓은데','크셔서','큽니다만','크신','길었습니다','덮었어요','덮어용','큼','크니깐','헐렁해요','덮은','긴하고','박시하니','길','크고요',
 '박시한','크거나','길긴한데','큰데도','길더라고요','크더라','버핏','펑퍼짐해는','크다','헐렁해서','루즈핏','컸어용','헐렁한','빡시','긴','길어여','길었어요',
 '커요','펑퍼짐해서','넓네요','크네요','크면','와이드','가려져서','크다는','덮히고','긴것','길다고','컸을','오버','남는데','넓어서','컸어요','크다고','크게',
 '깁니다','길더라구요','헐렁거리','크더라고요','큰지','길어서','접고','덮이수선해서','컸습니다','컸지만','걷어','큼직하게','길엇어','잡아먹힌','박시하고','박시한데',
 '남는','컸으나','컸다','크더라구요','덮고도','크니','길었지만','여유핏','넓고','큽니','크네','접어야','기네','큰듯','크실','큽니당','넓긴한데',
 '길긴하네요','박시핏','긴거','넓우','덮음','여유롭게','클','펑퍼짐한','큰','길고','헐렁하게','큰것','아방핏','흘러내리는','덮고','크지']

# 중복제거
def sizedict_nodup(oneOfRawList):
    oneOfRawList = set(oneOfRawList)
    return list(oneOfRawList)

new_chest     = sizedict_nodup(new_chest)
new_small     = sizedict_nodup(new_small)
new_big       = sizedict_nodup(new_big)

# 리스트와 리스트 이름들을 합쳐서 딕셔너리 생성
categories_chest = {
    'Chest': new_chest,
    'Small': new_small,
    'Big': new_big
}

## 5) arm_df : categories_arm

import pandas as pd

# 사전 단어

new_sleeve = ['소매','팔도','손목','손','팔다리','팔목','팔','팔꿈치','팔이','요팔',
       '당팔','팔길이','소매통','팔기장']

new_small = ['작습니다','짧았어요','짧습니다','짧으면','짧','작았어요','짧다고','쪼여서','작았으면',
      '작아도','작으면','작음','작아진','좁다고',
      '짧아요','짧게','짧다는','좁다는','작네','작기는','달라붙는','작은데','짧은데','짧음',
      '작긴','짧네요','작게','작아서','작은','짧고','짧아','좁은','타이트','짧았지만','붙어서','짧긴',
      '작은것','작아요','짧은','짧아서','짧지만','좁아','좁고','작네요','작아','쪼이는','좁아요','크롭느낌',
      '크롭하','작고','숏','작다고','작다','좁아서','작','크롭','미니','크롭해',
      '길었으면','컸으면','좁게','짧지' ,'작지','크롭이','핏되',
      '짝','좁', '작았고','작긴한데','짧긴하지만','짧아여','짧은게','작긴하지만','작으니','작았습니다','짧거나','좁긴'
                 '작으니까','작은듯','끼긴','짧은듯','작았는데','조이는','끼네요','붙고','달라붙어서','붙음','붙어요',
                 '짧은것','짧은거','끼는것','끼는','작은게','작음에도','짧은것이','좁네요','길었어도']

new_big = ['커도','남아','롱','긴게','넓게','덮어줘서','가오리','오퍼핏','부해','크구요','길고요','어벙함','컸었는데','길지도','길긴하지만','크니까','가리고',
 '박시하게','컸어도','덮는데','컸던','오벼핏','덮습니다','크고','덮','벌룬','덮을','헐렁하고','길거나','오버핏','박시하지만','덮이고','길게','길구요','길거든요',
 '걷어서','박스큽니다이','컷습니','컸고','박시하네요','박스핏','헐렁','덮어요','크','큼직해서','오바핏','길기는','넓아','크길래','넓어','루즈클까봐','낭',
 '클거','컸음','넓음','길다는','큰편인덴','넓었으면','수선해서','박시','널럴하','짧았으면','커욤','오비핏','넉','덮어서','덮네요','길어요','크지만','어벙',
 '덮구','길긴','낙','길어용','길지','덮이는','큰데','큰거','널','커서','펑퍼짐','길어','넓어요','길었고','더크','덮히는','넓길었는데','긴데','박시하','넓긴',
 '크긴','긴데도','덥는','접어도','넓은','크진','컸는데','큼직하니','벌룬핏','긴듯','커용','클까','오버사이즈','깁니','큼지막해서','접으면','박시해','가려요','커여',
 '넓은데','크셔서','큽니다만','크신','길었습니다','덮었어요','덮어용','큼','크니깐','헐렁해요','덮은','긴하고','박시하니','길','크고요',
 '박시한','크거나','길긴한데','큰데도','길더라고요','크더라','버핏','펑퍼짐해는','크다','헐렁해서','루즈핏','컸어용','헐렁한','빡시','긴','길어여','길었어요',
 '커요','펑퍼짐해서','넓네요','크네요','크면','와이드','가려져서','크다는','덮히고','긴것','길다고','컸을','오버','남는데','넓어서','컸어요','크다고','크게',
 '깁니다','길더라구요','헐렁거리','크더라고요','큰지','길어서','접고','덮이수선해서','컸습니다','컸지만','걷어','큼직하게','길엇어','잡아먹힌','박시하고','박시한데',
 '남는','컸으나','컸다','크더라구요','덮고도','크니','길었지만','여유핏','넓고','큽니','크네','접어야','기네','큰듯','크실','큽니당','넓긴한데',
 '길긴하네요','박시핏','긴거','넓우','덮음','여유롭게','클','펑퍼짐한','큰','길고','헐렁하게','큰것','아방핏','흘러내리는','덮고','크지']

# 중복제거
def sizedict_nodup(oneOfRawList):
    oneOfRawList = set(oneOfRawList)
    return list(oneOfRawList)

new_sleeve       = sizedict_nodup(new_sleeve)
new_small     = sizedict_nodup(new_small)
new_big       = sizedict_nodup(new_big)

# 리스트와 리스트 이름들을 합쳐서 딕셔너리 생성
categories_sleeve = {
    'Sleeve': new_sleeve,
    'Small': new_small,
    'Big': new_big
}

# 단어별 키워드 지정

## 1) total : wordSizDictotal

wordSizDictotal = {}

# categories 딕셔너리의 각 키와 값에 대해 반복
for size_key, words_list in categories_total.items():
    # 각 사이즈(key)에 대응하는 단어들을 반복하면서 wordSizDictotal에 추가
    for j in range(len(words_list)):
        wordSizDictotal[words_list[j]] = size_key

print(wordSizDictotal)


## 2) length : wordSizDiclength

wordSizDiclength= {}

# categories 딕셔너리의 각 키와 값에 대해 반복
for size_key, words_list in categories_length.items():
    # 각 사이즈(key)에 대응하는 단어들을 반복하면서 wordSizDiclength에 추가
    for j in range(len(words_list)):
        wordSizDiclength[words_list[j]] = size_key

print(wordSizDiclength)


## 3) shoulder : wordSizDicshoulder

wordSizDicshoulder = {}

# categories 딕셔너리의 각 키와 값에 대해 반복
for size_key, words_list in categories_shoulder.items():
    # 각 사이즈(key)에 대응하는 단어들을 반복하면서 wordSizDicshoulder에 추가
    for j in range(len(words_list)):
        wordSizDicshoulder[words_list[j]] = size_key

print(wordSizDicshoulder)


## 4) chest : wordSizDicchest

wordSizDicchest = {}

# categories 딕셔너리의 각 키와 값에 대해 반복
for size_key, words_list in categories_chest.items():
    # 각 사이즈(key)에 대응하는 단어들을 반복하면서 wordSizDicchest에 추가
    for j in range(len(words_list)):
        wordSizDicchest[words_list[j]] = size_key

print(wordSizDicchest)


## 5) sleeve : wordSizDicsleeve

wordSizDicsleeve = {}

# categories 딕셔너리의 각 키와 값에 대해 반복
for size_key, words_list in categories_sleeve.items():
    # 각 사이즈(key)에 대응하는 단어들을 반복하면서 wordSizDicsleeve에 추가
    for j in range(len(words_list)):
        wordSizDicsleeve[words_list[j]] = size_key

print(wordSizDicsleeve)


# 초기값 설정

## 리뷰 텍스트 -> 리스트 (전체)

new_df_catereviewlist = new_df["cate_review"].tolist()

len(new_df_catereviewlist)

## 카테고리별 df 컬럼명 설정 (전체)

sizes_total = ['Total', 'Small', 'Big']
sizes_length = ['Length', 'Small', 'Big']
sizes_shoulder = ['Shoulder', 'Small', 'Big']
sizes_chest = ['Chest', 'Small', 'Big']
sizes_sleeve = ['Sleeve', 'Small', 'Big']

### 1) total : df_sizedict_check_total

import pandas as pd
reviewIndex = []
for i in range(0,len(new_df_catereviewlist)):
  reviewIndex.append(str(i))
# 0으로 이루어진 단어 행렬 생성 (초기값)
sizedictCheck_total = pd.DataFrame(0.0, index=reviewIndex, columns=sizes_total)
print(sizedictCheck_total)

# 각 문서에서 단어가 등장할 때마다 행렬 값 업데이트
for i in range(len(new_df_catereviewlist)):
    sentence = new_df_catereviewlist[i]
    words = sentence.split(" ")
    for word in words:
        if word != '' and word in wordSizDictotal:
            sizedictCheck_total.loc[str(i), wordSizDictotal[word]] += 1.0
        else:
            pass  # wordSizDictotal에 해당 단어가 없으면 pass

print(sizedictCheck_total)

df_sizedict_check_total = pd.DataFrame(sizedictCheck_total)

### 2) length : df_sizedict_check_length

import pandas as pd
reviewIndex = []
for i in range(0,len(new_df_catereviewlist)):
  reviewIndex.append(str(i))
# 0으로 이루어진 단어 행렬 생성 (초기값)
sizedictCheck_length = pd.DataFrame(0.0, index=reviewIndex, columns=sizes_length)
print(sizedictCheck_length)

# 각 문서에서 단어가 등장할 때마다 행렬 값 업데이트
for i in range(len(new_df_catereviewlist)):
    sentence = new_df_catereviewlist[i]
    words = sentence.split(" ")
    for word in words:
        if word != '' and word in wordSizDiclength:
            sizedictCheck_length.loc[str(i), wordSizDiclength[word]] += 1.0
        else:
            pass  # wordSizDiclength에 해당 단어가 없으면 pass

print(sizedictCheck_length)

df_sizedict_check_length = pd.DataFrame(sizedictCheck_length)

### 3) shoulder : df_sizedict_check_shoulder

import pandas as pd
reviewIndex = []
for i in range(0,len(new_df_catereviewlist)):
  reviewIndex.append(str(i))
# 0으로 이루어진 단어 행렬 생성 (초기값)
sizedictCheck_shoulder = pd.DataFrame(0.0, index=reviewIndex, columns=sizes_shoulder)
print(sizedictCheck_shoulder)

# 각 문서에서 단어가 등장할 때마다 행렬 값 업데이트
for i in range(len(new_df_catereviewlist)):
    sentence = new_df_catereviewlist[i]
    words = sentence.split(" ")
    for word in words:
        if word != '' and word in wordSizDicshoulder:
            sizedictCheck_shoulder.loc[str(i), wordSizDicshoulder[word]] += 1.0
        else:
            pass  # wordSizDicshoulder에 해당 단어가 없으면 pass

print(sizedictCheck_shoulder)

df_sizedict_check_shoulder = pd.DataFrame(sizedictCheck_shoulder)

### 4) chest : df_sizedict_check_chest

import pandas as pd
reviewIndex = []
for i in range(0,len(new_df_catereviewlist)):
  reviewIndex.append(str(i))
# 0으로 이루어진 단어 행렬 생성 (초기값)
sizedictCheck_chest = pd.DataFrame(0.0, index=reviewIndex, columns=sizes_chest)
print(sizedictCheck_chest)

# 각 문서에서 단어가 등장할 때마다 행렬 값 업데이트
for i in range(len(new_df_catereviewlist)):
    sentence = new_df_catereviewlist[i]
    words = sentence.split(" ")
    for word in words:
        if word != '' and word in wordSizDicchest:
            sizedictCheck_chest.loc[str(i), wordSizDicchest[word]] += 1.0
        else:
            pass  # wordSizDicchest에 해당 단어가 없으면 pass

print(sizedictCheck_chest)

df_sizedict_check_chest = pd.DataFrame(sizedictCheck_chest)

### 5) sleeve : df_sizedict_check_sleeve

import pandas as pd
reviewIndex = []
for i in range(0,len(new_df_catereviewlist)):
  reviewIndex.append(str(i))
# 0으로 이루어진 단어 행렬 생성 (초기값)
sizedictCheck_sleeve = pd.DataFrame(0.0, index=reviewIndex, columns=sizes_sleeve)
print(sizedictCheck_sleeve)

# 각 문서에서 단어가 등장할 때마다 행렬 값 업데이트
for i in range(len(new_df_catereviewlist)):
    sentence = new_df_catereviewlist[i]
    words = sentence.split(" ")
    for word in words:
        if word != '' and word in wordSizDicsleeve:
            sizedictCheck_sleeve.loc[str(i), wordSizDicsleeve[word]] += 1.0
        else:
            pass  # wordSizDicsleeve에 해당 단어가 없으면 pass

print(sizedictCheck_sleeve)

df_sizedict_check_sleeve = pd.DataFrame(sizedictCheck_sleeve)

# 5개 카테고리 -> big/small 판단

## 1) total : df_sizedict_check_total

# float형을 int형으로 변환

# 새로운 열 total_big과 total_small 생성
df_sizedict_check_total['total_big'] = df_sizedict_check_total.apply(lambda row: 1 if row['Total'] > 0 and row['Big'] > 0 else 0, axis=1)
df_sizedict_check_total['total_small'] = df_sizedict_check_total.apply(lambda row: 1 if row['Total'] > 0 and row['Small'] > 0 else 0, axis=1)

# 기존의 total, big, small 열 삭제
df_sizedict_check_total.drop(columns=['Total', 'Big', 'Small'], inplace=True)

df_sizedict_check_total.head(100)

## 2) length : df_sizedict_check_length

# 새로운 열 total_big과 total_small 생성
df_sizedict_check_length['length_big'] = df_sizedict_check_length.apply(lambda row: 1 if row['Length'] > 0 and row['Big'] > 0 else 0, axis=1)
df_sizedict_check_length['length_small'] = df_sizedict_check_length.apply(lambda row: 1 if row['Length'] > 0 and row['Small'] > 0 else 0, axis=1)

# 기존의 total, big, small 열 삭제
df_sizedict_check_length.drop(columns=['Length', 'Big', 'Small'], inplace=True)

df_sizedict_check_length.head(100)

## 3) shoulder : df_sizedict_check_shoulder

# 새로운 열 total_big과 total_small 생성
df_sizedict_check_shoulder['shoulder_big'] = df_sizedict_check_shoulder.apply(lambda row: 1 if row['Shoulder'] > 0 and row['Big'] > 0 else 0, axis=1)
df_sizedict_check_shoulder['shoulder_small'] = df_sizedict_check_shoulder.apply(lambda row: 1 if row['Shoulder'] > 0 and row['Small'] > 0 else 0, axis=1)

# 기존의 total, big, small 열 삭제
df_sizedict_check_shoulder.drop(columns=['Shoulder', 'Big', 'Small'], inplace=True)

df_sizedict_check_shoulder.head(100)

## 4) chest : df_sizedict_check_chest

# 새로운 열 total_big과 total_small 생성
df_sizedict_check_chest['chest_big'] = df_sizedict_check_chest.apply(lambda row: 1 if row['Chest'] > 0 and row['Big'] > 0 else 0, axis=1)
df_sizedict_check_chest['chest_small'] = df_sizedict_check_chest.apply(lambda row: 1 if row['Chest'] > 0 and row['Small'] > 0 else 0, axis=1)

# 기존의 total, big, small 열 삭제
df_sizedict_check_chest.drop(columns=['Chest', 'Big', 'Small'], inplace=True)

df_sizedict_check_chest.head(100)

## 5) sleeve : df_sizedict_check_sleeve

# 새로운 열 total_big과 total_small 생성
df_sizedict_check_sleeve['sleeve_big'] = df_sizedict_check_sleeve.apply(lambda row: 1 if row['Sleeve'] > 0 and row['Big'] > 0 else 0, axis=1)
df_sizedict_check_sleeve['sleeve_small'] = df_sizedict_check_sleeve.apply(lambda row: 1 if row['Sleeve'] > 0 and row['Small'] > 0 else 0, axis=1)

# 기존의 total, big, small 열 삭제
df_sizedict_check_sleeve.drop(columns=['Sleeve', 'Big', 'Small'], inplace=True)

df_sizedict_check_sleeve.head(100)

# 병합

- cate_review null 값을 제거한 new_df와 카테고리별 5개 데이터프레임 병합

dfs = [new_df.reset_index(drop=True), df_sizedict_check_total.reset_index(drop=True), df_sizedict_check_length.reset_index(drop=True),
       df_sizedict_check_shoulder.reset_index(drop=True), df_sizedict_check_chest.reset_index(drop=True),
       df_sizedict_check_sleeve.reset_index(drop=True)]

merged_df = pd.concat(dfs, axis=1)

merged_df.shape

merged_df.shape, len(merged_df['goodsNo'].unique()), len(merged_df['review'].unique())

merged_df_copy = merged_df.copy()

# 최종 병합

import pandas as pd

# 특정 열 삭제 및 review, id 열을 기준으로 합치는 함수 정의
def merge_rows(df, key_columns, columns_to_delete):
    # 특정 열 삭제
    df = df.drop(columns=columns_to_delete)

    # 그룹별로 처리
    def custom_agg(series):
        # 값이 모두 같은지 확인
        if series.nunique() == 1:
            return series.iloc[0]
        else:
            return series.iloc[0]

    # 그룹별로 각 열에 대해 custom_agg 함수 적용
    merged_df = df.groupby(key_columns).agg(custom_agg).reset_index()

    return merged_df

# 함수 사용
result_df_4 = merge_rows(merged_df_copy, ['review', 'id', 'goodsNo'], ['cate_review'])

# 결과 출력
print(result_df_4.shape)
print(len(result_df_4['goodsNo'].unique()))


# csv 파일 저장
result_df_4.to_csv('result_df_4.csv', index=False)
